extract custom template functions and export the FuncMap
